requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
numpy==2.3.2
ta==0.11.0
yfinance==0.2.65
google-api-python-client==2.178.0
google-auth==2.40.3
python-dotenv==1.1.1
